// dashboard/src/api/models/creditNoteStatusEnum.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `draft` - Draft
 * `issued` - Issued
 * `voided` - Voided
 */
export type CreditNoteStatusEnum =
  (typeof CreditNoteStatusEnum)[keyof typeof CreditNoteStatusEnum];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreditNoteStatusEnum = {
  draft: "draft",
  issued: "issued",
  voided: "voided",
} as const;